added bordering of depth and comment

+++ minimax/minimax_mock.go
@@ -10,10 +10,10 @@ type StateMock struct {
 	mock.Mock
 }
 
-func (m *StateMock) Eval() float64 {
+func (m *StateMock) Eval() int {
 	args := m.Called()
 	fmt.Println("eval ", args.Get(0))
-	return args.Get(0).(float64)
+	return args.Get(0).(int)
 }
 
 func (m *StateMock) GetChildren(isMaximizer bool) []State {
